Bugfix: ROI labelling
Fixed a bug where the wrong labels were chosen to name the mask files. Instead of the community.short column of the mmp.csv file, we now use the roi column to do so. This has the advantage that each roi has a unique identifier, which was not the case for community.short.